Mass update of Apps (#145)
* Mass update of Apps

* Bulk request support

* Improvements on download

* Add delete download state

+++ src/system/resources/app/delete_data.py
@@ -1,22 +1,33 @@
 import os
 
-from flask_restful import reqparse
+from flask import request
+from rubix_http.exceptions.exception import PreConditionException
 from rubix_http.resource import RubixResource
 
 from src.system.apps.base.installable_app import InstallableApp
 from src.system.resources.app.utils import get_app_from_service
+from src.system.resources.rest_schema.schema_delete_data import delete_data_attributes
+from src.system.utils.data_validation import validate_args
 from src.system.utils.file import delete_existing_folder
 
 
 class DeleteDataResource(RubixResource):
     @classmethod
     def post(cls):
-        parser = reqparse.RequestParser()
-        parser.add_argument('service', type=str, required=True)
-        args = parser.parse_args()
-        service: str = args['service'].upper()
-        app: InstallableApp = get_app_from_service(service)
-        stop: bool = app.stop()
-        backup_data: bool = app.backup_data()
-        deletion: bool = delete_existing_folder(os.path.join(app.get_global_dir(), 'data'))
-        return {'service': service, 'deletion': deletion, 'backup_data': backup_data, 'stop': stop}
+        args = request.get_json()
+        if not validate_args(args, delete_data_attributes):
+            raise PreConditionException('Invalid request.')
+        delete_data_res = []
+        for arg in args:
+            service: str = arg['service'].upper()
+            app: InstallableApp = get_app_from_service(service)
+            try:
+                res = {'service': service, 'deletion': False, 'backup_data': False, 'stop': False, 'error': ''}
+                stop: bool = app.stop()
+                backup_data: bool = app.backup_data()
+                deletion: bool = delete_existing_folder(os.path.join(app.get_global_dir(), 'data'))
+                res = {**res, 'deletion': deletion, 'backup_data': backup_data, 'stop': stop}
+            except Exception as e:
+                res = {**res, 'error': str(e)}
+            delete_data_res.append(res)
+        return delete_data_res
